Create README.md
We have to run the project on the server. It will display a web page, wherein the search bar we have to write the name of any city in the world.
Then it will show us the weather data of that particular city, as well as the data, will also save on the database(here I use SQLite). We can delete the data from the view_page as well as from the database by clicking the X button on the view_page.
If the City name already exists on the database, then we have to delete first the old one then again have to search.